Reverting 7083,7079.
I tested locally with gcc 4.3, but gcc 4.2 behaves differently and fails.
Review URL: http://codereview.chromium.org/14165

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7087 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/ssl_error_info.cc
@@ -7,6 +7,7 @@
 #include "base/string_util.h"
 #include "chrome/browser/cert_store.h"
 #include "chrome/common/l10n_util.h"
+#include "chrome/common/resource_bundle.h"
 #include "chrome/common/time_format.h"
 #include "net/base/cert_status_flags.h"
 #include "net/base/net_errors.h"
@@ -234,7 +235,7 @@ int SSLErrorInfo::GetErrorsForCertStatus(int cert_id,
 
   scoped_refptr<net::X509Certificate> cert = NULL;
   int count = 0;
-  for (size_t i = 0; i < arraysize(kErrorFlags); ++i) {
+  for (int i = 0; i < arraysize(kErrorFlags); ++i) {
     if (cert_status & kErrorFlags[i]) {
       count++;
       if (!cert.get()) {
